requirements.txt
tk
chess
Pillow
pandas
Wand